+++ src/universal/Normalizer/EastNormalizerInterface.php
@@ -25,6 +25,9 @@ declare(strict_types=1);
 namespace Teknoo\East\Foundation\Normalizer;
 
 /**
+ * Interface to define normalizer, able to normalize via a serializer, like the Symfony's serializer, object following
+ * the East philosophy and implementing Teknoo\East\Foundation\Normalizer\Object\NormalizableInterface.
+ *
  * @copyright   Copyright (c) 2009-2017 Richard Déloge (richarddeloge@gmail.com)
  *
  * @link        http://teknoo.software/east Project website