+++ src/universal/Processor/ProcessorRecipeInterface.php
@@ -27,6 +27,9 @@ namespace Teknoo\East\Foundation\Processor;
 use Teknoo\East\Foundation\Recipe\RecipeInterface;
 
 /**
+ * Since 0.0.8, the Processor behavior is implementing into a Recipe. This interface allow to identify via a full
+ * qualified classname this recipe in the DI, and it's priority.
+ *
  * @copyright   Copyright (c) 2009-2017 Richard Déloge (richarddeloge@gmail.com)
  *
  * @link        http://teknoo.software/east Project website